Create atrix Framework for Analysis: Building a Matrix from Advanced Projects (I)
This raw dataset is part of a Framework Global - Matrix that enables a comprehensive and holistic assessment, simplifying comparisons and guiding decision-makers in selecting the alternatives that offer the greatest value in terms of financial return, social and environmental impact, and regulatory compliance. To streamline the analysis and comparison of indicators, the author utilized categories linked to Sustainable Competitiveness and the World Business Council for Sustainable Development (WBCSD). As a result, the indicators contribute to the functionality of the matrix, facilitating the identification of key projects related to the desired area and enabling future application of the data in sustainability-focused initiatives.